requirements.txt
hdbscan==0.8.29
numpy==1.24.3
opencv_python==4.7.0.72
Pillow==9.5.0
PyYAML==6.0
scikit_learn==1.2.2
